refactor(opal): split PageLoader into a layout
