+++ web/lib/opal/src/components/index.ts
@@ -113,9 +113,7 @@ export {
 /* Loader */
 export {
   OnyxLoader,
-  PageLoader,
   type OnyxLoaderProps,
-  type PageLoaderProps,
 } from "@opal/components/loader/components";
 
 /* Pagination */
