+++ web/lib/opal/src/components/loader/README.md
@@ -1,17 +1,15 @@
-# Loader
+# OnyxLoader
 
-**Import:** `import { OnyxLoader, PageLoader } from "@opal/components";`
+**Import:** `import { OnyxLoader } from "@opal/components";`
 
-Onyx-branded loading states. Animation timing and geometry follow the Onyx UI Library design.
+The Onyx-branded loading mark. Animation timing and geometry follow the Onyx UI Library design.
 
-- **`OnyxLoader`** renders the animated mark alone: the Onyx octagon outline and diamond logo crossfade while rotating a full turn on a 2s loop. Both layers use `currentColor` (default `border-02`), so the mark adapts to the surrounding theme. `size` prop in pixels, default 64 with a ~2.5px stroke that scales. Respects `prefers-reduced-motion` by holding the static outline.
-- **`PageLoader`** is the full-page loading state: the mark centered with a label beneath (`text` prop, `string | RichStr` for `markdown()` opt-in, default "Loading …"). Use for page and route-level loading. For an inline or section-level loader without a label, use `OnyxLoader` directly.
+`OnyxLoader` renders the animated mark alone: the Onyx octagon outline and diamond logo crossfade while rotating a full turn on a 2s loop. Both layers use `currentColor` (default `border-02`), so the mark adapts to the surrounding theme. `size` prop in pixels, default 64 with a ~2.5px stroke that scales. Respects `prefers-reduced-motion` by holding the static outline.
+
+For a full-page loading state with a centered label, use `PageLoader` from `@opal/layouts`.
 
 ```tsx
-// Route-level
-if (isLoading) return <PageLoader />;
-
-// Inline
+// Inline / section-level
 <OnyxLoader size={24} />
 ```
