+++ web/lib/opal/src/components/loader/components.tsx
@@ -1,6 +1,4 @@
 import "@opal/components/loader/styles.css";
-import type { RichStr } from "@opal/types";
-import { Text } from "@opal/components";
 
 // ---------------------------------------------------------------------------
 // OnyxLoader
@@ -85,29 +83,4 @@ function OnyxLoader({ size = 64 }: OnyxLoaderProps) {
   );
 }
 
-// ---------------------------------------------------------------------------
-// PageLoader
-// ---------------------------------------------------------------------------
-
-interface PageLoaderProps {
-  /** Label beneath the mark, markdown() opt-in. @default "Loading …" */
-  text?: string | RichStr;
-}
-
-/**
- * Full-page loading state: the animated Onyx mark with a label, centered in
- * the available space. Use for page/route-level loading. For an inline or
- * section-level loader without a label, use `OnyxLoader` directly.
- */
-function PageLoader({ text = "Loading …" }: PageLoaderProps) {
-  return (
-    <div className="flex h-full min-h-[60vh] w-full flex-col items-center justify-center gap-3 p-5">
-      <OnyxLoader />
-      <Text font="main-ui-muted" color="text-03">
-        {text}
-      </Text>
-    </div>
-  );
-}
-
-export { OnyxLoader, PageLoader, type OnyxLoaderProps, type PageLoaderProps };
+export { OnyxLoader, type OnyxLoaderProps };
