+++ web/lib/opal/src/layouts/index.ts
@@ -101,3 +101,9 @@ export {
   ConfirmationModalLayout,
   type ConfirmationModalProps,
 } from "@opal/layouts/modal/components";
+
+/* PageLoader */
+export {
+  PageLoader,
+  type PageLoaderProps,
+} from "@opal/layouts/page-loader/components";
